zram: move init_done check to the beginning of disksize_store
If we set disksize when disksize has been set, we will get the following
error report: "write error: Cannot allocate memory". This is because
disksize_store fails at zram_meta_alloc. This is not what we expect.
To solve this problem, this patch moves init_done check to the beginning
of disksize_store.

Signed-off-by: Geliang Tang <geliangtang@163.com>
Signed-off-by: Rohit Vashisht <rohitv88cf@gmail.com>