fs: proc: Squash Android Verity Flags
* proc: Remove verifiedbootstate flag from /proc/cmdline
Userspace parses this and sets the ro.boot.verifiedbootstate prop
according to the value that this flag has. When ro.boot.verifiedbootstate
is not 'green', SafetyNet is tripped and fails the CTS test.
Hide verifiedbootstate from /proc/cmdline in order to fix the failed
SafetyNet CTS check.

* proc: Remove additional SafetyNet flags from /proc/cmdline
SafetyNet checks androidboot.veritymode in Nougat, so remove it.
Additionally, remove androidboot.enable_dm_verity and androidboot.secboot
in case SafetyNet will check them in the future.
Signed-off-by: Sultanxda <sultanxda@gmail.com>

* md: dm android verity always green

Signed-off-by: Rohit Vashisht <rohitv88cf@gmail.com>